requirements.txt
Flask==1.1.2
google-cloud-storage==1.31.0
google-auth==1.21.0
six==1.15.0